chore(gitignore): don't save virtio iso
VirtIO is a standardized framework that provides a set of
drivers and transport mechanisms for virtual machines (VMs).
It is designed to enhance the performance and efficiency of I/O
(input/output) operations in virtualized environments.

Before packer build, virtio-win.iso should be download
to be installed in windows qemu box.